requirements.txt
psycopg2
python-dotenv
Flask
Flask-SQLAlchemy
Flask-Migrate
black
mypy